Refactored the code as db schema got changed on 12th may
project_id is removed from activity_master.
project_user_mapping table got dropped, which impacted to the operation getAllProjectsByUser.

+++ zieta/src/main/java/com/zietaproj/zieta/response/ProjectDetailsByUserModel.java
@@ -15,7 +15,7 @@ public class ProjectDetailsByUserModel {
 	String projectName;
 	long projectType;
 	String OrgNode;
-	short projectStatus;
+	long projectStatus;
 	long clientId;
 
 	String projectManager;
